Change PT_integer_T from unsigned long long int to int.

+++ Problem-Types.h
@@ -11,7 +11,7 @@ This source code is developed based on the BSF skeleton (https://github.com/leon
 #include "Problem-Parameters.h"		// Problem Parameters 
 //=========================== Problem Types =========================
 typedef float						PT_float_T;
-typedef unsigned long long int		PT_integer_T;
+typedef int                 		PT_integer_T;
 typedef PT_float_T					PT_vector_T[PP_MAX_N];
 typedef PT_vector_T					PT_matrix_T[PP_MAX_M];
 //typedef PT_float_T				PT_point_T[PP_MAX_N];
